Keep video view registration internal
