+++ lib/src/track/video_track_view_registration.dart
@@ -1,0 +1,33 @@
+// Copyright 2024 LiveKit, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import 'package:flutter/material.dart';
+
+import 'package:meta/meta.dart';
+
+import '../types/other.dart';
+
+@internal
+class VideoTrackViewRegistration {
+  /// The widget key used by adaptive stream to find this view's render context.
+  final GlobalKey key = GlobalKey();
+
+  /// The pixel density used to convert this view's logical size to physical
+  /// pixels when computing adaptive-stream dimensions.
+  AdaptiveStreamPixelDensity pixelDensity;
+
+  VideoTrackViewRegistration({
+    this.pixelDensity = AdaptiveStreamPixelDensity.auto,
+  });
+}
